New job for reloading client list details
Because a new field was added to client list which isn't in old data, there is now a way for users to request that all the details in their client list (company name and status) are reloaded from the database. Doesn't call the API, just gets profiles from Redis. Users will have to call this to get statuses added.

Signed-off-by: Brian Evans <ebrian101@gmail.com>