Added backend support for recent filings
Took quite a bit of design work to get this set up. The data is being fetched correctly from Redis, but needs to be set to be loaded correctly when new filings or clients are added. The readme in dataPlugins/recentFilings explains the design. Its quite an efficient way of getting filings in a specific period of time.

Signed-off-by: Brian Evans <ebrian101@gmail.com>