Add events from streaming API to Redis Stream
Each org gets its own Redis Stream named org:orgId:notifications, which gets populated by events on the streaming API if they are for companies in the org client list. The plan for the future is to listen on this Redis stream for email/web notifications, and also allow users to view a history of notifications. Testing for the filing stream doesn't seem to be working, not sure why. Company profile is definitely working though.

Signed-off-by: Brian Evans <ebrian101@gmail.com>