Improved worker for listening on streams
Used Fetch API instead of https.Request.

Shouldn't throw errors anymore, if there is an internal server error, it will keep retrying to connect. If there is an error with the request it will exit because it needs to be fixed in the source code. More detailed logging, including response headers, status code and in the event of an error it logs the response body. Should lead to a more reliable worker with less downtime. Also re-enabled the worker job in Github actions because it should stop in time now, but this will need to be tested when the event stream is quiet.

Signed-off-by: Brian Evans <ebrian101@gmail.com>