Refactor send notification function
Extracted logic to send a notification out to backend-shared, so that notifications can easily be sent by both the server and the worker.
Also fixed a recently introduced bug in the navbar for the current user.
Also made the service worker public and uncompiled for more consistency in dev/prod.
Added an onClick handler to notifications.

Signed-off-by: Brian Evans <ebrian101@gmail.com>