Changed recent filings data format
Instead of having the data grouped in an object, it's now a flat array and its up to the client to group the data however the user wishes. This paves the way for grouping by attributes other than filingType such as date and company.
Happened to find a bug in the ErrorAlert at the same time, so fixed that.

Signed-off-by: Brian Evans <ebrian101@gmail.com>