More stream worker error handling
The streaming API worker was shutting down the container when Companies House ended the connection every night. Hopefully now it should catch and emit the error so that it can be handled properly. This is hard to test without actually being disconnected by Companies House. I've also removed batching from the worker logs so that they are sent in event of a shutdown.

Signed-off-by: Brian Evans <ebrian101@gmail.com>