Fix bug in format filing description
This bug was found in production by a user. Some filing descriptions don't have an enum, which means it calls replace on an undefined variable when formatting. To solve this, a check is now done that only tries to format if the description has an enum. Logs a warning if not and doesn't throw error.

Signed-off-by: Brian Evans <ebrian101@gmail.com>