Warn user about incomplete filing data
When a new client list is uploaded, there is a rate limit for loading the filing history, so they are not always loaded immediately. The user is now warned while this background loading is going on, so that they are aware that the recent filings page may not have complete data.
Also freed some RAM by changing the background command for the REPL.

Signed-off-by: Brian Evans <ebrian101@gmail.com>